requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.3
django-toolbelt==0.0.1
gunicorn==19.1.1
mixpanel-py==4.0.0
psycopg2==2.5.4
pytz==2014.10
requests==2.5.1
static3==0.5.1
waitress==0.8.9
whitenoise==1.0.6
